+++ test/Sentry/Tracing/EventHandlerTest.php
@@ -14,7 +14,7 @@ class EventHandlerTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $handler = new EventHandler($this->app, $this->app->make(BacktraceHelper::class), []);
+        $handler = new EventHandler([], $this->app->make(BacktraceHelper::class));
 
         /** @noinspection PhpUndefinedMethodInspection */
         $handler->thisIsNotAHandlerAndShouldThrowAnException();
@@ -36,7 +36,7 @@ class EventHandlerTest extends TestCase
 
     private function tryAllEventHandlerMethods(array $methods): void
     {
-        $handler = new EventHandler($this->app, $this->app->make(BacktraceHelper::class), []);
+        $handler = new EventHandler([], $this->app->make(BacktraceHelper::class));
 
         $methods = array_map(static function ($method) {
             return "{$method}Handler";